setup script füer switcher2

+++ setup_swi.sh
@@ -23,14 +23,36 @@ else
 fi
 
 echo "Files fuer mosquitto gemacht"
+echo " "
+echo "kopiere scripts für switcher"
 
 #
-cp -n -v shell_scripts/switcher2.sh /etc/init.d
-cp -n -v shell_scripts/swserver2.sh /etc/init.d
-update-rc.d switcher2.sh defaults
-update-rc.d swserver2.sh defaults
+file="/etc/init.d/switcher2.sh"
+if [ -f "$file" ]
+then
+	echo "$file bereits vorhanden."
+else
+	echo "$file not found, also kopiere"
+    cp -n -v shell_scripts/switcher2.sh /etc/init.d
+fi
+
+file="/etc/init.d/swserver2.sh"
+if [ -f "$file" ]
+then
+	echo "$file bereits vorhanden."
+else
+	echo "$file not found, also kopiere"
+    cp -n -v shell_scripts/swserver2.sh /etc/init.d
+fi
+
 chmod 755 /etc/init.d/switcher2.sh
 chmod 755 /etc/init.d/swserver2.sh
+
+update-rc.d switcher2.sh defaults
+update-rc.d swserver2.sh defaults
+
 #
+echo "Files fuer switcher gemacht"
+
